refactor(diff-panel): extract display item builder
